functions/save_exoplanets/requirements.txt

functions-framework
google-cloud-firestore